Move shareable search service code to Microsoft.PackageManagement.Search.Web (#979)
