+++ src/Microsoft.PackageManagement.Search.Web/Support/ApiExceptionFilterAttribute.cs
@@ -1,13 +1,13 @@
 ﻿// Copyright (c) .NET Foundation. All rights reserved.
 // Licensed under the Apache License, Version 2.0. See License.txt in the project root for license information.
 
+using System.Net;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.AspNetCore.Mvc.Filters;
 using NuGet.Services.AzureSearch;
 using NuGet.Services.AzureSearch.SearchService;
-using System.Net;
 
-namespace NuGet.Services.SearchService
+namespace Microsoft.PackageManagement.Search.Web
 {
     public class ApiExceptionFilterAttribute : ExceptionFilterAttribute
     {
@@ -21,7 +21,6 @@ namespace NuGet.Services.SearchService
                         StatusCode = (int)HttpStatusCode.ServiceUnavailable
                     };
                     break;
-
                 case InvalidSearchRequestException isre:
                     context.Result = new JsonResult(new ErrorResponse(isre.Message))
                     {
