+++ src/NuGet.Services.SearchService.Core/NuGet.Services.SearchService.Core.csproj
@@ -8,22 +8,9 @@
   </PropertyGroup>
 
   <ItemGroup>
-    <PackageReference Include="Microsoft.ApplicationInsights.AspNetCore" Version="2.16.0" />
-    <PackageReference Include="Microsoft.AspNetCore.Mvc.NewtonsoftJson" Version="3.1.10" />
-
-    <!--
-    This reference is used to "lift" a transitive dependency of Microsoft.ApplicationInsights.AspNetCore to a version
-    that is accepted by Azure DevOps Component Governance, which scans the project.assets.json file. At runtime, the
-    real version this package are actually higher and is defined by the shared ASP.NET Core SDK installed on the
-    hosting machine.
-    -->
-    <PackageReference Include="Microsoft.AspNetCore.Http" Version="2.2.2" />
+    <ProjectReference Include="..\Microsoft.PackageManagement.Search.Web\Microsoft.PackageManagement.Search.Web.csproj" />
   </ItemGroup>
 
-  <ItemGroup>
-    <ProjectReference Include="..\NuGet.Services.AzureSearch\NuGet.Services.AzureSearch.csproj" />
-  </ItemGroup>
-  
   <Import Project="$(NuGetBuildExtensions)" Condition="'$(NuGetBuildExtensions)' != '' And Exists('$(NuGetBuildExtensions)')" />
-  
+
 </Project>
