+++ src/NuGet.Services.SearchService.Core/Startup.cs
@@ -1,6 +1,11 @@
 // Copyright (c) .NET Foundation. All rights reserved.
 // Licensed under the Apache License, Version 2.0. See License.txt in the project root for license information.
 
+using System;
+using System.Collections.Generic;
+using System.Net;
+using System.Net.Http;
+using System.Text.Json.Serialization;
 using Autofac;
 using Microsoft.ApplicationInsights;
 using Microsoft.ApplicationInsights.Extensibility;
@@ -10,25 +15,16 @@ using Microsoft.AspNetCore.Mvc;
 using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Hosting;
+using Microsoft.PackageManagement.Search.Web;
 using NuGet.Services.AzureSearch;
 using NuGet.Services.AzureSearch.SearchService;
-using NuGet.Services.Configuration;
-using NuGet.Services.KeyVault;
 using NuGet.Services.Logging;
 using NuGet.Services.SearchService.Controllers;
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Net;
-using System.Net.Http;
-using System.Text.Json.Serialization;
 
 namespace NuGet.Services.SearchService
 {
     public class Startup
     {
-        public const string EnvironmentVariablePrefix = "APPSETTING_";
-        private const string ControllerSuffix = "Controller";
         private const string ConfigurationSectionName = "SearchService";
 
         public Startup(IConfiguration configuration)
@@ -40,7 +36,7 @@ namespace NuGet.Services.SearchService
 
         public void ConfigureServices(IServiceCollection services)
         {
-            var refreshableConfig = GetSecretInjectedConfiguration(Configuration);
+            var refreshableConfig = StartupHelper.GetSecretInjectedConfiguration(Configuration);
             Configuration = refreshableConfig.Root;
             services.AddSingleton(refreshableConfig.SecretReaderFactory);
 
@@ -68,11 +64,11 @@ namespace NuGet.Services.SearchService
             });
             services.AddSingleton<ITelemetryInitializer>(new KnownOperationNameEnricher(new[]
             {
-                GetOperationName<SearchController>(HttpMethod.Get, nameof(SearchController.AutocompleteAsync)),
-                GetOperationName<SearchController>(HttpMethod.Get, nameof(SearchController.IndexAsync)),
-                GetOperationName<SearchController>(HttpMethod.Get, nameof(SearchController.GetStatusAsync)),
-                GetOperationName<SearchController>(HttpMethod.Get, nameof(SearchController.V2SearchAsync)),
-                GetOperationName<SearchController>(HttpMethod.Get, nameof(SearchController.V3SearchAsync)),
+                StartupHelper.GetOperationName<SearchController>(HttpMethod.Get, nameof(SearchController.AutocompleteAsync)),
+                StartupHelper.GetOperationName<SearchController>(HttpMethod.Get, nameof(SearchController.IndexAsync)),
+                StartupHelper.GetOperationName<SearchController>(HttpMethod.Get, nameof(SearchController.GetStatusAsync)),
+                StartupHelper.GetOperationName<SearchController>(HttpMethod.Get, nameof(SearchController.V2SearchAsync)),
+                StartupHelper.GetOperationName<SearchController>(HttpMethod.Get, nameof(SearchController.V3SearchAsync)),
             }));
             services.AddApplicationInsightsTelemetryProcessor<SearchRequestTelemetryProcessor>();
             services.AddSingleton<TelemetryClient>();
@@ -101,92 +97,13 @@ namespace NuGet.Services.SearchService
 
         public void Configure(IApplicationBuilder app, IWebHostEnvironment env)
         {
-            if (env.IsDevelopment())
-            {
-                app.UseDeveloperExceptionPage();
-            }
-
-            app.UseRouting();
+            StartupHelper.Configure(app, env);
 
             app.UseCors(cors => cors
                 .AllowAnyOrigin()
                 .WithHeaders("Content-Type", "If-Match", "If-Modified-Since", "If-None-Match", "If-Unmodified-Since", "Accept-Encoding")
                 .WithMethods("GET", "HEAD", "OPTIONS")
                 .WithExposedHeaders("Content-Type", "Content-Length", "Last-Modified", "Transfer-Encoding", "ETag", "Date", "Vary", "Server", "X-Hit", "X-CorrelationId"));
-
-            app.UseHsts();
-
-            app.Use(async (context, next) =>
-            {
-                context.Response.Headers.Add("X-Content-Type-Options", "nosniff");
-                await next();
-            });
-
-            app.UseEndpoints(endpoints =>
-            {
-                endpoints.MapControllers();
-            });
-        }
-
-        private static RefreshableConfiguration GetSecretInjectedConfiguration(IConfigurationRoot uninjectedConfiguration)
-        {
-            // Initialize KeyVault integration.
-            var secretReaderFactory = new ConfigurationRootSecretReaderFactory(uninjectedConfiguration);
-            var refreshSecretReaderSettings = new RefreshableSecretReaderSettings();
-            var refreshingSecretReaderFactory = new RefreshableSecretReaderFactory(secretReaderFactory, refreshSecretReaderSettings);
-            var secretReader = refreshingSecretReaderFactory.CreateSecretReader();
-            var secretInjector = refreshingSecretReaderFactory.CreateSecretInjector(secretReader);
-
-            // Attempt to inject secrets into all of the configuration strings.
-            foreach (var pair in uninjectedConfiguration.AsEnumerable())
-            {
-                if (!string.IsNullOrWhiteSpace(pair.Value))
-                {
-                    // We can synchronously wait here because we are outside of the request context. It's not great
-                    // but we need to fetch the initial secrets for the cache before activating any controllers or
-                    // asking DI for configuration.
-                    secretInjector.InjectAsync(pair.Value).Wait();
-                }
-            }
-
-            // Reload the configuration with secret injection enabled. This is used by the application.
-            var injectedBuilder = new ConfigurationBuilder()
-                .AddInjectedJsonFile("appsettings.json", secretInjector)
-                .AddInjectedJsonFile("appsettings.Development.json", secretInjector)
-                .AddInjectedEnvironmentVariables(EnvironmentVariablePrefix, secretInjector);
-            var injectedConfiguration = injectedBuilder.Build();
-
-            // Now disable all secrets loads from a non-refresh path. Refresh will be called periodically from a
-            // background thread. Foreground (request) threads MUST use the cache otherwise there will be a deadlock.
-            refreshSecretReaderSettings.BlockUncachedReads = true;
-
-            return new RefreshableConfiguration
-            {
-                SecretReaderFactory = refreshingSecretReaderFactory,
-                Root = injectedConfiguration,
-            };
-        }
-
-        private static string GetControllerName<T>() where T : ControllerBase
-        {
-            var typeName = typeof(T).Name;
-            if (typeName.EndsWith(ControllerSuffix, StringComparison.Ordinal))
-            {
-                return typeName.Substring(0, typeName.Length - ControllerSuffix.Length);
-            }
-
-            throw new ArgumentException($"The controller type name must end with '{ControllerSuffix}'.");
-        }
-
-        private static string GetOperationName<T>(HttpMethod verb, string actionName) where T : ControllerBase
-        {
-            return $"{verb} {GetControllerName<T>()}/{actionName}";
-        }
-
-        private class RefreshableConfiguration
-        {
-            public IRefreshableSecretReaderFactory SecretReaderFactory { get; set; }
-            public IConfigurationRoot Root { get; set; }
         }
     }
 }
