+++ tests/NuGet.Services.SearchService.Core.Tests/Support/ApiExceptionFilterAttributeFacts.cs
@@ -9,6 +9,7 @@ using Microsoft.AspNetCore.Mvc;
 using Microsoft.AspNetCore.Mvc.Abstractions;
 using Microsoft.AspNetCore.Mvc.Filters;
 using Microsoft.AspNetCore.Routing;
+using Microsoft.PackageManagement.Search.Web;
 using Moq;
 using NuGet.Services.AzureSearch;
 using NuGet.Services.AzureSearch.SearchService;
